Various changes
Updated manifest. Now enables the app to flip in landscape mode.
Updated GameActivity. Timer is now stopped properly on finish.
Minor update in GameLogFragment. Very small change to a field name.
Updated GameLog. Now correctly displaying logs with the oldest at the
very bottom. Was previously displaying with the oldest at the bottom,
but with the period order with the oldest period at the top.